# Add a cache in front of every Rival accelerator