--- a/src/syntax/generators.rkt
+++ b/src/syntax/generators.rkt
@@ -5,6 +5,7 @@
          ffi/unsafe)
 
 (require "../core/rival.rkt"
+         "../config.rkt"
          "types.rkt")
 
 (provide from-rival
@@ -24,14 +25,23 @@
 
 ; ----------------------- RIVAL GENERATOR ---------------------------
 
+(define/reset caches '()
+  (lambda ()
+    (for ([cache (caches)])
+      (hash-clear! cache))))
+
 (define-generator ((from-rival) spec ctx)
   (define compiler (make-real-compiler (list spec) (list ctx)))
   (define fail ((representation-bf->repr (context-repr ctx)) +nan.bf))
+  (define cache (make-hash))
+  (caches (cons cache (caches)))
   (lambda pt
-    (define-values (_ exs) (real-apply compiler (list->vector pt)))
-    (if exs
-        (first exs)
-        fail)))
+    (hash-ref! cache pt
+               (lambda ()
+                 (define-values (_ exs) (real-apply compiler (list->vector pt)))
+                 (if exs
+                     (first exs)
+                     fail)))))
 
 ; ----------------------- FFI GENERATOR -----------------------------