--- a/src/syntax/generators.rkt
+++ b/src/syntax/generators.rkt
@@ -34,7 +34,7 @@
 (define-generator ((from-rival #:cache? [cache? #t]) spec ctx)
   (define compiler (make-real-compiler (list spec) (list ctx)))
   (define fail ((representation-bf->repr (context-repr ctx)) +nan.bf))
-  (define (compute pt)
+  (define (compute . pt)
     (define-values (_ exs) (real-apply compiler (list->vector pt)))
     (if exs
         (first exs)
@@ -43,7 +43,7 @@
     [cache?
      (define cache (make-hash))
      (caches (cons cache (caches)))
-     (lambda pt (hash-ref! cache pt (lambda () (compute pt))))]
+     (lambda pt (hash-ref! cache pt (lambda () (apply compute pt))))]
     [else compute]))
 
 ; ----------------------- FFI GENERATOR -----------------------------