Rival "accelerators" are just platform operators defined with `from-rival`. These are _super duper slow_. So slow! We've got examples of them taking a benchmark from 1s → 40s. Some investigation of this for the `grow-libm` branch by @JonasRegehr revealed that we are frequently evaluating an accelerator with the _same exact input_ multiple times. This PR thus adds a cache in every `from-rival`-generated operator. One tricky bit is making sure we clear these between benchmarks.

For example, across `tutorial.rkt`, we seem to do 2.6M evaluations of a `sum5` accelerator in the `slow-rival` branch. After deduplicating I think there are 416k, which is a 6x reduction, and unsurprisingly the total runtime goes from 77s to 18s (about a 4x reduction).

I'm not 100% sure we should actually merge this—ideally we'd just make Rival fast enough that caching isn't worth it—but let's discuss Monday.